==> 1_raw_data_processing.ipynb <==
import pandas as pd
import os
import numpy as np
from IPython.display import display
import pycountry
# pd.set_option('display.max_rows', )

datasets_path = 'dataset/'

# Load the csv dataset
raw_df = pd.read_csv(os.path.join(datasets_path, 'athlete_events.csv'))

# Visualize df info
raw_df.info()

# First 5 rows to make sure of data types
display(raw_df.head(5))

import re

# Find the -1 -2 suffixes from names and teams and take them off
p = '(-\d+)'
raw_df.Team = raw_df.Team.str.replace(p, '', case=False, regex=True)

raw_df.Name = raw_df.Name.str.replace(p, '', case=False, regex=True)

# Cast to appropriate data types

# Sex
raw_df.Sex = raw_df.Sex.astype(pd.CategoricalDtype(categories=['M', 'F']))
raw_df.Age = raw_df.Age.astype('Int8')
raw_df.Height = np.floor(pd.to_numeric(raw_df.Height, errors='coerce')).astype('Int16')
raw_df.Weight = raw_df.Weight.astype('float32')
raw_df.NOC = raw_df.NOC.astype('category')
raw_df.Games = raw_df.Games.astype('category')
raw_df.Year = raw_df.Year.astype('uint16')
raw_df.Season = raw_df.Season.astype('category')
raw_df.City = raw_df.City.astype('category')
raw_df.Sport = raw_df.Sport.astype('category')
raw_df.Event = raw_df.Event.astype('category')
raw_df.Medal = raw_df.Medal.astype(pd.CategoricalDtype(categories=['Gold', 'Silver', 'Bronze'], ordered=True))
raw_df.Team = raw_df.Team.astype('category')

# Find duplicates
dups = raw_df.duplicated(subset=['ID', 'Name', 'Sex', 'Team', 'Games', 'Year', 'Season', 'City', 'Sport', 'Event', 'Medal'], keep=False)
dups_df = raw_df[dups]
display(dups_df)

raw_df = raw_df.drop_duplicates(subset=['ID', 'Name', 'Sex', 'Team', 'Games',
                                        'Year', 'Season', 'City', 'Sport',
                                        'Event', 'Medal'])

raw_df['Medal'] = raw_df['Medal'].astype(pd.CategoricalDtype(categories=['Gold', 'Silver', 'Bronze', 'NoMedal'], ordered=True))
raw_df['Medal'] = raw_df['Medal'].fillna(value='NoMedal')


display(raw_df.head(5))
raw_df.info()
# Save to a parquet silver table
raw_df.to_parquet(os.path.join(datasets_path, 'silver_df.parquet'))



noc_df = pd.read_csv(os.path.join(datasets_path, 'noc_regions.csv'))

# Visualize df info
noc_df.info()
noc_df.head(10)



# Check for duplicates
dups = noc_df.duplicated(subset=['NOC', 'region'], keep=False)
print(dups.unique())

# noc_df.to_parquet(os.path.join(datasets_path, 'silver_noc.parquet'))

query_country_df = noc_df.dropna(subset=['region'])


countries_alpha_3 = []
for region, code in zip(query_country_df.region, query_country_df.NOC):
    try:
        country_returned = pycountry.countries.lookup(region)
        alpha3_print = country_returned.alpha_3
    except LookupError:
        # Try to find a match with the code
        try:
            country_returned = pycountry.countries.lookup(code)
            alpha3_print = country_returned.alpha_3
        except LookupError:
            # Try to find a fuzzy match
            try:
                country_returned = pycountry.countries.search_fuzzy(region)
                if country_returned == 1:
                    alpha3_print = country_returned[0].alpha_3
                else:
                    print(f'Region {region} ({code}), multiple match found: {country_returned}')
                    alpha3_print = ''
            except LookupError:
                print(f'Region {region} ({code}), no match found')
                alpha3_print = ''
    countries_alpha_3.append(alpha3_print)
query_country_df = query_country_df.assign(alpha_3 = countries_alpha_3)
query_country_df = query_country_df.set_index('NOC')
display(query_country_df)

query_country_df['alpha_3'].loc['AHO'] = 'CUW'
query_country_df['alpha_3'].loc['ANT'] = 'ATG'
query_country_df['alpha_3'].loc['BRU'] = 'BRN'
query_country_df['alpha_3'].loc['CGO'] = 'COG'
query_country_df['alpha_3'].loc['EUN'] = 'RUS'
query_country_df['alpha_3'].loc['IOA'] = ''
query_country_df['alpha_3'].loc['IRI'] = 'IRN'
query_country_df['alpha_3'].loc['KOS'] = ''
query_country_df['alpha_3'].loc['PLE'] = 'PSE'
query_country_df['alpha_3'].loc['SKN'] = 'KNA'
query_country_df['alpha_3'].loc['UAR'] = 'SYR'
query_country_df['alpha_3'].loc['URS'] = 'SUN'
query_country_df['alpha_3'].loc['VIN'] = 'VCT'
query_country_df['alpha_3'].loc['WIF'] = 'TTO'
query_country_df['alpha_3'].loc['ISV'] = 'VIR'
query_country_df['alpha_3'].loc['HKG'] = 'HKG'
display(query_country_df.head(3))

# Get country full names
country_names = []
for alpha_3 in query_country_df.alpha_3:
    print(f'{alpha_3}')
    if not alpha_3:
        print(f'No country code')
        country_names.append('None')
    else:
        try:
            country_returned = pycountry.countries.get(alpha_3= alpha_3)
            if not country_returned:
                country_returned = pycountry.historic_countries.get(alpha_3= alpha_3)
            if country_returned:
                country_names.append(country_returned.name)
            else:
                print(f'{alpha_3} not found')
                country_names.append('None')
        except LookupError:
            print(f'{alpha_3} not found')
            country_names.append('None')


query_country_df = query_country_df.assign(iso_names=country_names)
query_country_df = query_country_df.reset_index()
query_country_df = query_country_df.drop(columns=['notes'])
query_country_df.to_parquet(os.path.join('dataset', 'iso_countries.parquet'))
display(query_country_df)
