==> 4_exploratory_data.ipynb <==
import pandas as pd
import os
from IPython.display import display
from pandasql import sqldf
import numpy as np
dataset_path = 'dataset'
pysqldf = lambda q: sqldf(q, globals())


silver_df = pd.read_parquet(os.path.join(dataset_path, 'silver_df.parquet'))


# Like for 3_descriptive_analysis
# Extract the number of different athletes for each NOC per Year for Summer Olympics

count_distinct_athletes = f" SELECT NOC, Year, COUNT(DISTINCT ID) as distinct_athletes_no" \
                f" FROM silver_df" \
                f" WHERE Season='Summer' " \
                 f" GROUP BY NOC, Year"
gold_distinct_athletes_df = pysqldf(count_distinct_athletes)



# Validate the previous query by looking at the number of distinct athletes for team USA
# In the last 3 editions (2016, 2012, 2008)

# Results should be:
# - 2016: 554 (263 men and 291 women) in 33 sports
# - 2012: 530 (262 men and 268 women) in 31 sports
# - 2008: 588 (306 men and 282 women) in 32 sports

# display(gold_distinct_athletes_df.loc[gold_distinct_athletes_df.NOC == 'USA'])
# Query result: 555 (+1), 530, 588.
# The team USA website reports 558 participants for Rio 2016


# Extract the number of competitions entries for each NOC per Year for Summer Olympics

count_events_partecipations = f" SELECT NOC, Year, COUNT(1) as events_partecipations_no" \
                f" FROM silver_df" \
                f" WHERE Season='Summer' " \
                 f" GROUP BY NOC, Year"
gold_events_partecipations_df = pysqldf(count_events_partecipations)

# display(gold_events_partecipations_df.loc[gold_events_partecipations_df.NOC == 'USA'])


# Extract number of medals by type for each NOC


subquery_count_winning = f" SELECT NOC, Sport, Event, Year, Season, Games, Team, Medal" \
                f" FROM silver_df" \
                f" WHERE Medal IS NOT NULL AND Season='Summer' " \
                 f" GROUP BY NOC, Sport, Event, Year, Season, Games, Team, Medal" \
                f" ORDER BY NOC"
medalists_df = pysqldf(subquery_count_winning)



# NOC, year, and different medals
noc_medals_df = pd.pivot_table(medalists_df,
                               index=['NOC', 'Year'],
                               columns='Medal',
                               aggfunc='count',
                               values='Event',
                               fill_value=0,
                               margins=True,
                               margins_name='TotalPartecipants')

noc_medals_df = noc_medals_df.drop(('TotalPartecipants',''))
noc_medals_df.columns.name = ''
noc_medals_df.reset_index(inplace=True)
display(noc_medals_df)


# gold_partecipants_df contains NOC, Year, events_partecipations_no, distinct_athletes_no, Bronze, Silver, Gold, Total
gold_partecipants_df = gold_events_partecipations_df.merge(gold_distinct_athletes_df, on=['NOC', 'Year'], how='left')
gold_partecipants_df = gold_partecipants_df.merge(noc_medals_df, on=['NOC', 'Year'], how='left')
gold_partecipants_df = gold_partecipants_df.fillna(value=0)
gold_partecipants_df.Gold = gold_partecipants_df.Gold.astype(int)
gold_partecipants_df.Silver = gold_partecipants_df.Silver.astype(int)
gold_partecipants_df.Bronze = gold_partecipants_df.Bronze.astype(int)
gold_partecipants_df.TotalPartecipants = gold_partecipants_df.TotalPartecipants.astype(int)
gold_partecipants_df['TotalMedals'] = gold_partecipants_df.Gold + gold_partecipants_df.Silver + gold_partecipants_df.Bronze
gold_partecipants_df = gold_partecipants_df.loc[gold_partecipants_df.Year > 1960]
display(gold_partecipants_df)

# Read NOC table
noc_df = pd.read_parquet(os.path.join(dataset_path, 'iso_countries.parquet'))
gold_partecipants_df = gold_partecipants_df.merge(noc_df, on=['NOC'], how='left')
gold_partecipants_df.fillna('Not Found')
display(gold_partecipants_df)

# Regions not found
display(gold_partecipants_df.loc[gold_partecipants_df.region == 'Not Found'])


# Column Temporary aide in population search
gold_partecipants_df['population_help'] = gold_partecipants_df.NOC


# Replacement countries needed for population fill
# 'GDR' 'GER'
# 'ROT'
# 'VNM' 'VIE'
# 'EUN' 'RUS'
# 'YUG' 'SRB'
# 'IOA'
# 'URS' 'RUS'
# 'FRG' 'GER'
# 'YAR' 'YEM'
# 'SCG' 'SRB'
# 'TCH' 'CZE'
# 'YMD' 'YEM'
gold_partecipants_df['population_help'] = gold_partecipants_df['population_help'].replace(
    to_replace=['GDR', 'VNM', 'EUN', 'YUG', 'URS', 'FRG', 'YAR', 'SCG', 'TCH', 'YMD'],
    value=['GER', 'VIE', 'RUS','SRB', 'RUS',  'GER', 'YEM', 'SRB', 'CZE', 'YEM']
)

display(gold_partecipants_df)


# Load population data
# From https://data.worldbank.org/indicator/SP.POP.TOTL
# and https://population.un.org/wpp/Download/Standard/Population/
population_df = pd.read_parquet(os.path.join(dataset_path, 'populations.parquet'))
population_df.info()

population_df = population_df.set_index(['Country Code', 'Year'])
display(population_df.head(5))

# Fill population
test_pop = population_df.loc[(gold_partecipants_df.NOC, gold_partecipants_df.Year.astype(int)), 'Population'].reset_index()
gold_partecipants_df = gold_partecipants_df.merge(test_pop, left_on=['population_help', 'Year'], right_on=['Country Code', 'Year'], how='left')


# Cleaning up columns
gold_partecipants_df = gold_partecipants_df.drop(columns=['population_help', 'Country Code'])


# CHECKPOINT: my full db aggregated by country with no nulls
# Countries with no known population (like refugees olympic team) are discarded

gold_partecipants_df.to_parquet(os.path.join(dataset_path, 'gold_partecipants_zero_pop.parquet'))

# Countries with no population
no_population_df = gold_partecipants_df[gold_partecipants_df['Population'].isna()]
print(f'Number of rows with no population: {len(no_population_df)}')
display(no_population_df)

gold_partecipants_df = gold_partecipants_df.dropna(subset=['Population'])
gold_partecipants_df.to_parquet(os.path.join(dataset_path, 'gold_partecipants.parquet'))


# Entriees with no partecipants
print(f'Number of rows with no partecipants: {len(gold_partecipants_df[gold_partecipants_df.TotalPartecipants == 0])}')



display(gold_partecipants_df.info())

# Metrics for countries those with one medal or more
gold_metrics_partecipants_df = gold_partecipants_df[gold_partecipants_df.TotalMedals > 0]
display(gold_metrics_partecipants_df)

gold_metrics_partecipants_df = gold_metrics_partecipants_df.assign(
    PopulationPerMedal_thousands = np.divide(np.divide(gold_metrics_partecipants_df.Population, 1000), gold_metrics_partecipants_df.TotalMedals) )
gold_metrics_partecipants_df = gold_metrics_partecipants_df.assign(
    EventPartecipationPerMedal = np.divide(gold_metrics_partecipants_df.events_partecipations_no, gold_metrics_partecipants_df.TotalMedals) )
gold_metrics_partecipants_df = gold_metrics_partecipants_df.assign(
    AthletePerMedal = np.divide(gold_metrics_partecipants_df.distinct_athletes_no, gold_metrics_partecipants_df.TotalMedals) )
gold_metrics_partecipants_df = gold_metrics_partecipants_df.assign(
    AthletePerEventPartecipation = np.divide(gold_metrics_partecipants_df.distinct_athletes_no, gold_metrics_partecipants_df.events_partecipations_no) )

display(gold_metrics_partecipants_df.head(5))


gold_metrics_partecipants_df.drop(columns='Year').describe()

print('Gold:')
display(gold_partecipants_df.sort_values(by='Gold', ascending=False).head(5))
print('Silver:')
display(gold_partecipants_df.sort_values(by='Silver', ascending=False).head(5))
print('Bronze:')
display(gold_partecipants_df.sort_values(by='Bronze', ascending=False).head(5))
print('Total Medals:')
display(gold_partecipants_df.sort_values(by='TotalMedals', ascending=False).head(5))


gold_metrics_partecipants_df.to_parquet(os.path.join('dataset','gold_countries_year_metrics.parquet'))
