==> 4b_athletes_heigh_performance.ipynb <==

import pandas as pd
import os
from IPython.display import display
from pandasql import sqldf
import numpy as np
from scipy import stats
import plotly.express as px
dataset_path = 'dataset'
pysqldf = lambda q: sqldf(q, globals())


silver_df = pd.read_parquet(os.path.join(dataset_path, 'silver_df.parquet'))



query_df = silver_df.loc[silver_df.Year > 1960]
query_df['Medal'] = query_df['Medal'].astype(pd.CategoricalDtype(categories=['Gold', 'Silver', 'Bronze', 'NoMedal'], ordered=True))
query_df['Medal'] = query_df['Medal'].fillna(value='NoMedal')




# NOC, year, and different medals
noc_height_df = pd.pivot_table(query_df,
                               index=['Height', 'Sex'],
                               columns='Medal',
                               aggfunc='count',
                               values='Event',
                               fill_value=0,
                               margins=True,
                               margins_name='TotalPartecipants',
                               observed=True)
noc_height_df = noc_height_df.drop(('TotalPartecipants',''))
noc_height_df.columns.name = ''
noc_height_df.reset_index(inplace=True)
noc_height_df.Height = noc_height_df.Height.astype(int)
display(noc_height_df.describe())

noc_height_df.to_parquet(os.path.join('dataset','noc_height.parquet'))

threshold = noc_height_df.TotalPartecipants.quantile(0.15)
noc_height_df = noc_height_df[noc_height_df.TotalPartecipants>threshold]
noc_height_df['TotalMedals'] = noc_height_df.Gold + noc_height_df.Silver + noc_height_df.Bronze
noc_height_df['MedalPerPartecipant'] = np.divide(noc_height_df.TotalMedals, noc_height_df.TotalPartecipants)
print(f"Filter out heights that have less than {threshold:.1f} datapoints")


fig = px.scatter(x=noc_height_df.Height, y=noc_height_df['MedalPerPartecipant'], color=noc_height_df.Sex)
# fig.update_layout(title='Height probability density based on sex and year',
#                    yaxis_title='PDF (-)',
#                    xaxis_title='Height (cm)',
#                   legend={'title': 'Sex:'})
fig.show()

male_df = noc_height_df.loc[noc_height_df.Sex == "M"]
female_df = noc_height_df.loc[noc_height_df.Sex == "F"]


print('Male:')
male_df.corr(method='pearson')

print('Female:')
female_df.corr(method='pearson')

print(stats.pearsonr(female_df.Height, female_df.MedalPerPartecipant))

print(stats.pearsonr(male_df.Height, male_df.MedalPerPartecipant))
