==> 8_presentation.ipynb <==
# Countries Olympics performance evaluation
## Data analysis project for Italy National Olympic Committee
### October 2021
#### Andrea Biasioli

## Goal of the analysis
The analysis relies on the SportsStats dataset to study country-wide patterns in terms of medal earning.

The goal is to understand which countries are currently performing better and which countries are on the rise.

The analysis will delve on how the country is faring respect to others of similar size.

Additional information provided include an analysis on whether different NOCs rely on single high-talent individuals competing in more editions/events, and which parameters correlate better with winning medals.


# Question 1
## Which countries are currently the most successful (in terms of number of medals) and which ones are on the rise?

- It can be helpful to know if similarly-sized countries are faring better or worse
- It can also be helpful to see which countries are consistently performing well (maybe they established a good funding/coaching system)




# Question 2
## Which countries relies mostly on single-athletes performance to obtain their medals?

Some countries performance will be reliant on single individuals. This is not too promising for their future, as when the athlete retires the country performance will likely worsen.

We want to make sure that our NOC is not one of them.





# Question 3
## Is there any connection between the height and the expected performance?

- Sometimes young athletes are selected based on their technical skills, and not their physical potential. If a relationship is found between height and performance, maybe the recruiting process could be different and encourage physical-based down selection.
- Height may be relevant in more recent editions but not in earlier ones. Is the importance changing in time?
- The analysis should be differentiated based on sex
